This relates to the string cutter that I made before.
This will take a file in and convert it into a string with the
file_get_contents feature from PHP and then with that, the limit
is applied and the message length is detected as well and is cut
in a way that the words are not cut in half.
When fileReader.php has done its job, fileWriter.php will take over
with the file and will use file_put_contents to insert the phrase
cut from fileReader.php and then use FILE_APPEND to append the file
and not overwrite it. It will also use PHP_EOL to insert two lines
down so that it is not written on the same line and it can be seen
that it has done its job.